+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#322] Prevent "resource name may not be empty" on dogu delete
+- [#323] Requeue dogu deletions to be able to resolve SA dependency problems
 
 ## [v3.27.1] - 2026-07-31
 ### Added